refactor: перенос сортировки фильмов по жанру в FilmDbStorage.java

+++ src/main/java/ru/yandex/practicum/filmorate/dao/impl/FilmDbStorage.java
@@ -132,24 +132,30 @@ public class FilmDbStorage implements FilmStorage {
     }
 
     @Override
-    public List<Film> getTopFilms(Integer count, String year) {
+    public List<Film> getTopFilms(Integer count, Integer genreId, String year) {
         log.debug("Получен запрос вывести список популярных фильмов");
 
-        if (year == null) {
+        String baseSql = "SELECT f.*, "
+                + "rm.name AS rating_name "
+                + "FROM films f "
+                + "JOIN rating_MPA rm ON rm.ID = f.rating_id "
+                + "LEFT JOIN film_genres fg ON fg.film_id = f.id "
+                + "WHERE (YEAR(f.release_date) = ?) %s (fg.genre_id = ?) "
+                + "ORDER BY likes DESC LIMIT ?";
+
+        if (year == null && genreId == null) {
             String sql = "SELECT f.*, "
                     + "rm.name AS rating_name "
                     + "FROM films f "
                     + "JOIN rating_MPA rm ON rm.ID = f.rating_id "
                     + "ORDER BY likes DESC LIMIT ?";
             return jdbcTemplate.query(sql, mapper::makeFilm, count);
+        } else if (year == null || genreId == null) {
+            String nonStrictSql = String.format(baseSql, "OR");
+            return jdbcTemplate.query(nonStrictSql, mapper::makeFilm, year, genreId, count);
         } else {
-            String sql = "SELECT f.*, "
-                    + "rm.name AS rating_name "
-                    + "FROM films f "
-                    + "JOIN rating_MPA rm ON rm.ID = f.rating_id "
-                    + "WHERE YEAR(f.release_date) = ?"
-                    + "ORDER BY likes DESC LIMIT ?";
-            return jdbcTemplate.query(sql, mapper::makeFilm, Integer.parseInt(year), count);
+            String strictSql = String.format(baseSql, "AND");
+            return jdbcTemplate.query(strictSql, mapper::makeFilm, Integer.parseInt(year), genreId, count);
         }
     }
 }
